refactor: menu button structure and attributes
This patch will create component folder `Menu` and separate the buttons
into its own `MenuButton` component file.

It also refactors the active button class handler, instead of a ternary
to use a html data-attribute.

The `Icon` parameter required by the previous button component was
refactored to allow its usage as `icon` in lowercase.
